refactor uncompressed codec to its own source file (#805)
